Create a LabStatus standalone class.
