+++ packages/application/src/index.ts
@@ -6,7 +6,7 @@ import '../style/index.css';
 
 export { JupyterFrontEnd, JupyterFrontEndPlugin } from './frontend';
 
-export { ILabStatus, JupyterLab } from './lab';
+export { JupyterLab } from './lab';
 
 export { ILayoutRestorer, LayoutRestorer } from './layoutrestorer';
 
@@ -15,3 +15,5 @@ export { IMimeDocumentTracker } from './mimerenderers';
 export { IRouter, Router } from './router';
 
 export { ILabShell, LabShell } from './shell';
+
+export { ILabStatus } from './status';
